+++ packages/gatsby-remark-katex/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.0"></a>
+
+# [3.0.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-katex/compare/gatsby-remark-katex@2.0.6...gatsby-remark-katex@3.0.0) (2018-11-15)
+
+**Note:** Version bump only for package gatsby-remark-katex
+
 <a name="2.0.6"></a>
 
 ## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-katex/compare/gatsby-remark-katex@2.0.5...gatsby-remark-katex@2.0.6) (2018-10-29)